Move schema validation out of workflows
