+++ src/test/example-file-test.ts
@@ -2,7 +2,7 @@ import * as fs from "fs";
 
 import { describe, it } from "./util/cappucino";
 import { expect } from "chai";
-import { FetchRemoteSchemas, LoadIfcxFile, SchemaValidationError } from "../ifcx-core/workflows";
+import { FetchRemoteSchemas, LoadIfcxFile } from "../ifcx-core/workflows";
 import { IfcxFile } from "../ifcx-core/schema-helper";
 
 let examplesFolderPath = "../examples"
